Move Qt utils to _qt folder
